src/Python/examples/modules/flask/src/requirements.txt
python-dotenv==0.17.1
Flask==2.0.1